requirements.txt
Django==3.2.6
django-mptt==0.11.0
django-notifications-hq==1.6.0
djangorestframework==3.12.2
pymarc==4.1.1
pytz==2021.1
serpy==0.3.1
tornado==6.1


tablib~=3.0.0
openpyxl~=3.0.7